End Back & Start Front

+++ backend/controllers/product.controller.js
@@ -37,9 +37,9 @@ export const updateProduct = async (req, res) => {
     const {id} = req.params;
   
     const product = req.body;
-  
+    
     if(!mongoose.Types.ObjectId.isValid(id)){
-      return res.status(404).send(`No product with this id: ${id}`);    
+        return res.status(404).json({success: false, message:"Invalid Product ID" });    
     }
   
     try {
@@ -54,11 +54,15 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => { // :id --> dynamic variable
     const {id} = req.params;
     
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({success: false, message:"Invalid Product ID" });    
+    }
+
     try{
       await Product.findByIdAndDelete(id);
       res.status(200).json({success: true, message: "Product successfully deleted"});
     } catch (error) {
       console.log("Error in deleting product:", error.message);
-      res.status(404).json({success: false, message:"Product Not Found"});
+      res.status(500).json({success: false, message:"Server Error"});
     }
   }
